# APIFinal/bin/Release/net6.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "APIFinal, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "paths": {
...
    "/api/Productos": {
      "get": {
        "tags": [
          "Productos"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "post": {
        "tags": [
          "Productos"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/ProductosModel"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "put": {
        "tags": [
          "Productos"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/ProductosModel"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Productos/{id}": {
      "get": {
        "tags": [
          "Productos"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Productos"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/api/UsuarioLogin/{id}": {
      "get": {
        "tags": [
          "UsuarioLogin"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "UsuarioLogin"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    }
  },
  "definitions": {
...
    "CategoriaModel": {
      "type": "object",
      "properties": {
        "categoriaId": {
          "format": "int32",
          "type": "integer"
        },
        "descripcion": {
          "maxLength": 100,
          "type": "string"
        }
      }
    },
...
    "InsumosModel": {
      "type": "object",
      "properties": {
        "insumoId": {
          "format": "int32",
          "type": "integer"
        },
        "nombre": {
          "maxLength": 100,
          "type": "string"
        },
        "descripcion": {
          "maxLength": 100,
          "type": "string"
        },
        "unidadId": {
          "format": "int32",
          "type": "integer"
        },
        "tipoUnidad": {
          "$ref": "#/definitions/TipoUnidadModel"
        }
      }
    },
...
    "PreparacionInsumoModel": {
      "type": "object",
      "properties": {
        "preparacioninsumoId": {
          "format": "int32",
          "type": "integer"
        },
        "cantidad": {
          "format": "int32",
          "type": "integer"
        },
        "insumoId": {
          "format": "int32",
          "type": "integer"
        },
        "insumos": {
          "$ref": "#/definitions/InsumosModel"
        }
      }
    },
    "PreparacionModel": {
      "type": "object",
      "properties": {
        "preparacionId": {
          "format": "int32",
          "type": "integer"
        },
        "insumoId": {
          "format": "int32",
          "type": "integer"
        },
        "preparacionInsumo": {
          "$ref": "#/definitions/PreparacionInsumoModel"
        }
      }
    },
    "ProductosModel": {
      "type": "object",
      "properties": {
        "productoId": {
          "format": "int32",
          "type": "integer"
        },
        "nombre": {
          "maxLength": 100,
          "type": "string"
        },
        "descripcion": {
          "maxLength": 200,
          "type": "string"
        },
        "precio": {
          "format": "double",
          "type": "number"
        },
        "categoriaId": {
          "format": "int32",
          "type": "integer"
        },
        "preparacionId": {
          "format": "int32",
          "type": "integer"
        },
        "categorias": {
          "$ref": "#/definitions/CategoriaModel"
        },
        "preparacion": {
          "$ref": "#/definitions/PreparacionModel"
        }
      }
    },
...
    "TipoUnidadModel": {
      "type": "object",
      "properties": {
        "unidadId": {
          "format": "int32",
          "type": "integer"
        },
        "nombreUnidad": {
          "maxLength": 65,
          "type": "string"
        }
      }
    },
...
  }
}